# docs/swagger.yaml
basePath: /
definitions:
...
  dto.AddOrderDto:
    properties:
      customerId:
        example: 1
        type: integer
      products:
        items:
          $ref: '#/definitions/dto.AddOrderProductDto'
        type: array
      totalAmount:
        example: 34.99
        type: number
    type: object
  dto.AddOrderProductDto:
    properties:
      price:
        example: 34.99
        type: number
      productId:
        example: 1
        type: integer
      quantity:
        example: 1
        type: integer
    type: object
...
  dto.GetCustomerResponseDto:
    properties:
      cpf:
        type: integer
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  dto.GetOrderResponseDto:
    properties:
      created_at:
        type: string
      customer:
        $ref: '#/definitions/dto.GetCustomerResponseDto'
      id:
        type: integer
      products:
        items:
          $ref: '#/definitions/dto.OrderProductDto'
        type: array
      status:
        items:
          $ref: '#/definitions/dto.GetOrderStatusResponseDto'
        type: array
      total_amount:
        type: number
    type: object
  dto.GetOrderStatusResponseDto:
    properties:
      created_at:
        type: string
      current_status:
        type: integer
      current_status_description:
        type: string
      id:
        type: integer
      order_id:
        type: integer
    type: object
  dto.GetOrdersResponseDto:
    properties:
      orders:
        items:
          $ref: '#/definitions/dto.GetOrderResponseDto'
        type: array
    type: object
...
  dto.OrderProductDto:
    properties:
      category:
        type: integer
      description:
        type: string
      image_link:
        type: string
      name:
        type: string
      price:
        type: number
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
  dto.UpdateOrderStatusRequestDto:
    properties:
      status:
        example: 1
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Api to manage a fast food restaurant
  title: Tc-Fiap-50
  version: "1.0"
paths:
...
  /v1/order:
    get:
      consumes:
      - application/json
      description: Get orders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetOrdersResponseDto'
      summary: Get orders
      tags:
      - Order
    post:
      consumes:
      - application/json
      description: Add order
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AddOrderDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.GetOrderResponseDto'
      summary: Add order
      tags:
      - Order
  /v1/order/{orderId}:
    get:
      consumes:
      - application/json
      description: Get order
      parameters:
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetOrderResponseDto'
      summary: Get order
      tags:
      - Order
  /v1/order/{orderId}/status:
    get:
      consumes:
      - application/json
      description: Get order status
      parameters:
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetOrderStatusResponseDto'
      summary: Get order status
      tags:
      - Order
    put:
      consumes:
      - application/json
      description: Update order status
      parameters:
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: integer
      - description: Status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateOrderStatusRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Update order status
      tags:
      - Order
...
swagger: "2.0"